api: tm: support different packet priority modes
Current TM spec associates a priority to a packet at TM queue and
expects all the levels to honour that priority globally in the TM
system. In some platforms, this is not supportable and a packet's
priority gets overridden when it passes through a fan-in node, it
is passing through. In other words, like assigning packet a
priority based on the TM queue it is enqueued on, every node it
is passing through overrides the packet priority with a fixed value.

So this patch adds flexibility for platforms to expose pkt priority
modes supported and also choose the mode per TM system among supported
modes at time of odp_tm_create().

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>