api: tm: support different packet priority modes
Current TM spec associates a priority to a packet at TM queue and
expects all the levels to honour that priority globally in the TM
system. In some platforms, this is not supportable and a packet's
priority gets overridden when it passes through a fan-in node, it
is passing through. In other words, like assigning packet a
priority based on the TM queue it is enqueued on, every node it
is passing through overrides the packet priority with a fixed value.

So this patch adds flexibility for platforms to expose pkt priority
modes supported and also choose the mode per TM system among supported
modes at time of odp_tm_create().

This patch also adds "max_wfq_groups_per_node" capability that is
expressed by a platform when the mode is ODP_TM_PKT_PRIO_MODE_OVERRIDE.
In ODP_TM_PKT_PRIO_MODE_OVERRIDE mode, WFQ group is formed by a group of
fan-in nodes that feed pkts of same priority and WFQ needs to be done
among these fan-in nodes.
This field expresses maximum number of WFQ groups can be formed at a node.
Some platforms have a limitation where there is a limit on how many WFQ
groups can be formed at a given parent tm node with max being max
priorities supported.

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>