api: tm: add option to enable packet mode in shaper
Similar to using packet/frame count in scheduling
algorithm and weight in units of number of packets,
add option to enable packet mode in shaper
by introducing a boolean flag in shaper profile params.
When packet mode is set to TRUE, commit/peak rate is
in PPS (packets per second) and commit/peak burst is in
Packets.

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>