api: proto_stats: introduce generic protocol stats framework
Add generic protocol stats framework where protocol stats object
supports basic operations like creation, lookup, access and destroy
of stats object is supported. Each object supports a set of
stats based on object params.

This generic framework is also useful when we have to associate stats
objects to fast path with minimal overhead as it is as close to HW
implementation as possible.

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>
Signed-off-by: Jerin Jacob <jerinj@marvell.com>